python/requirements.txt
flask
mysql-connector-python
flask-cors
pytz 
flask_talisman
python-dotenv
